tests/wpt/web-platform-tests/app-history/navigate/back-forward-multiple-frames.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<iframe id="i" src="/common/blank.html"></iframe>
<script>
promise_test(async t => {
  // Wait for after the load event so that the navigation doesn't get converted
  // into a replace navigation.
  await new Promise(resolve => window.onload = () => t.step_timeout(resolve, 0));
  // Step 1
  assert_equals(appHistory.entries().length, 1);
  assert_equals(i.contentWindow.appHistory.entries().length, 1);
  await appHistory.navigate("#top");
  // Step 2: iframe at initial entry, top on second entry
  assert_equals(appHistory.entries().length, 2);
  assert_equals(i.contentWindow.appHistory.entries().length, 1);
  await i.contentWindow.appHistory.navigate("#iframe");

  // Step 3: Both windows on second entry.
  assert_equals(appHistory.entries().length, 2);
  assert_equals(i.contentWindow.appHistory.entries().length, 2);
  assert_equals(appHistory.current.index, 1);
  assert_equals(i.contentWindow.appHistory.current.index, 1);

  // NOTE: the order of navigation in the two windows is not guaranteed, so
  // do a timeout after each promise resolves before verifying the result, to
  // ensure adequate time is given for all navigations to complete.

  // Going back in the iframe should go 3->2 (navigating iframe only)
  await i.contentWindow.appHistory.back().then(t.step_func(() => new Promise(resolve => t.step_timeout(resolve, 0))));
  assert_equals(appHistory.current.index, 1);
  assert_equals(i.contentWindow.appHistory.current.index, 0);

  // Going forward in iframe should go 2->3
  await i.contentWindow.appHistory.forward().then(t.step_func(() => new Promise(resolve => t.step_timeout(resolve, 0))));
  assert_equals(appHistory.current.index, 1);
  assert_equals(i.contentWindow.appHistory.current.index, 1);

  // Going back in top should go 3->1 (navigating both windows).
  await appHistory.back().then(t.step_func(() => new Promise(resolve => t.step_timeout(resolve, 0))));
  assert_equals(appHistory.current.index, 0);
  assert_equals(i.contentWindow.appHistory.current.index, 0);

  // Going forward in top should go 1->2 (navigating top only)
  await appHistory.forward().then(t.step_func(() => new Promise(resolve => t.step_timeout(resolve, 0))));
  assert_equals(appHistory.current.index, 1);
  assert_equals(i.contentWindow.appHistory.current.index, 0);

  // Going back in top should go 2->1
  await appHistory.back().then(t.step_func(() => new Promise(resolve => t.step_timeout(resolve, 0))));
  assert_equals(appHistory.current.index, 0);
  assert_equals(i.contentWindow.appHistory.current.index, 0);

  // Going forward in iframe should go 1->3 (navigating both windows)
  await i.contentWindow.appHistory.forward().then(t.step_func(() => new Promise(resolve => t.step_timeout(resolve, 0))));
  assert_equals(appHistory.current.index, 1);
  assert_equals(i.contentWindow.appHistory.current.index, 1);
}, "appHistory.back() and appHistory.forward() can navigate multiple frames");
</script>
